Added a new option for the generation of the projection mapping coordinate systems
The new option simplifies the fitting of projection coordinate systems: Instead of placing the three points anywhere, it instead
places them at the most extreme points of the target region (like it does for the source coordinates) and just optimizes within a
permutation of the three points.

Activate by specifying "only_permutation": 1 in the "fit_args": {} part of the configuration.