Added a "fallback" layer profile for use in unlayered regions.
In order to parameterize projections into unlayered regions, we need to be able to express that no layer profile is to be used at all.
Particularly, this came up with respect to cortico-thalamic connections.

I added to layer_profile_mixer.py the following behavior:
No matter what is configured, an additional layer profile will be added as a fallback for non-layered regions:

- name: profile_10
  relative_densities:
  - layers:
    - ALL
    value: 1.0

This "ALL" indicates that layers are to be disregarded and simply uniform density (1.0 everywhere) is to be applied.

Next, the method to predict the most likely layer profile is updated to always return said fallback profile if the target of a projection is non-layered.
It gets this information from the RegionMapper object (in parcellation/region_mapper.py), to which I added the method is_layered(...).
The RegionMapper in turn reads that information from the configuration file. To that end, a section "layered_regions" is now expected in the configuration under "BrainParcellation".
It contains a list of all regions that have layers, i.e that are to NOT use the fallback. If you don't intend to use the fallback at all, place here simply a copy of the "region_names" entry.

This breaks compatibility with configurations that do not have the entry!

I then updated the example configurations to use the new entry. Then I removed an older configuration that lacks the entry (default_template.json).